+++ PowerDeComposerTest/src/test/java/com/xbreeze/xml/test/RunPDCUnitTest.java
@@ -11,7 +11,7 @@ import static io.cucumber.junit.platform.engine.Constants.FILTER_TAGS_PROPERTY_N
 @Suite
 @IncludeEngines("cucumber")
 @SelectClasspathResource("com/xbreeze/xml/test")
-@ConfigurationParameter(key = PLUGIN_PROPERTY_NAME, value="pretty")
+@ConfigurationParameter(key = PLUGIN_PROPERTY_NAME, value="pretty,json:target/cucumber-reports/UnitTest.json")
 @ConfigurationParameter(key = FILTER_TAGS_PROPERTY_NAME, value="@Unit")
 public class RunPDCUnitTest {
 }